Add more advanced config variable parsing.
This commit allows configuration variables to appear anywhere in the
string. For example:

   Given URL path "/person/#{person_id}"

This allows construction of URLs based on variables and also JSON paths
based on variables, for example:

  And JSON Pointer "/#{email}/exists" is False

This commit does not change existing functionality but adds the ability
for more complex variable parsing. The variable prefixes stay the same
but are required to be wrapped in {}. For example:

    Environment variables prefixed by '${}'
    config.ini prefixed by '@{}'
    test case context prefixed by '#{}'